+++ .gitignore
@@ -1,1 +1,1 @@
-venv/
+backend/venv/
